+++ rust/worker/src/blockstore/arrow/mod.rs
@@ -1,7 +1,6 @@
 mod block;
 pub(crate) mod blockfile;
 pub(crate) mod flusher;
-pub(crate) mod iterator;
 pub(crate) mod provider;
 mod sparse_index;
 pub(crate) mod types;
